test: rename all authorization test cases
The tests have been renamed from ...IT to ...Test. This is the only change in this commit.
