+++ zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/authorization/permissions/DecisionEvaluationEvaluateAuthorizationTest.java
@@ -18,16 +18,14 @@ import io.camunda.client.protocol.rest.PermissionTypeEnum;
 import io.camunda.client.protocol.rest.ResourceTypeEnum;
 import io.camunda.zeebe.it.util.AuthorizationsUtil;
 import io.camunda.zeebe.it.util.AuthorizationsUtil.Permissions;
-import io.camunda.zeebe.model.bpmn.Bpmn;
-import io.camunda.zeebe.protocol.record.intent.ProcessInstanceIntent;
 import io.camunda.zeebe.qa.util.cluster.TestStandaloneBroker;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration.TestZeebe;
 import io.camunda.zeebe.test.util.junit.AutoCloseResources;
 import io.camunda.zeebe.test.util.junit.AutoCloseResources.AutoCloseResource;
-import io.camunda.zeebe.test.util.record.RecordingExporter;
 import io.camunda.zeebe.test.util.testcontainers.TestSearchContainers;
 import java.util.List;
+import java.util.Map;
 import java.util.UUID;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -38,15 +36,13 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 @AutoCloseResources
 @Testcontainers
 @ZeebeIntegration
-public class ProcessInstanceModificationModifyAuthorizationIT {
-  public static final String JOB_TYPE = "jobType";
-  public static final String SERVICE_TASK_ID = "serviceTask";
+public class DecisionEvaluationEvaluateAuthorizationTest {
 
   @Container
   private static final ElasticsearchContainer CONTAINER =
       TestSearchContainers.createDefeaultElasticsearchContainer();
 
-  private static final String PROCESS_ID = "processId";
+  private static final String DECISION_ID = "jedi_or_sith";
   private static AuthorizationsUtil authUtil;
   @AutoCloseResource private static CamundaClient defaultUserClient;
 
@@ -69,79 +65,68 @@ public class ProcessInstanceModificationModifyAuthorizationIT {
     authUtil.awaitUserExistsInElasticsearch(defaultUsername);
     defaultUserClient
         .newDeployResourceCommand()
-        .addProcessModel(
-            Bpmn.createExecutableProcess(PROCESS_ID)
-                .startEvent()
-                .serviceTask(SERVICE_TASK_ID, t -> t.zeebeJobType(JOB_TYPE))
-                .endEvent()
-                .done(),
-            "process.xml")
+        .addResourceFromClasspath("dmn/drg-force-user.dmn")
         .send()
         .join();
   }
 
   @Test
-  void shouldBeAuthorizedToModifyProcessWithDefaultUser() {
-    // given
-    final var processInstanceKey = createProcessInstance();
-    final var serviceTaskKey = getServiceTaskKey(processInstanceKey);
-
-    // when then
+  void shouldBeAuthorizedToEvaluateDecisionWithDefaultUser() {
+    // when
     final var response =
         defaultUserClient
-            .newModifyProcessInstanceCommand(processInstanceKey)
-            .terminateElement(serviceTaskKey)
+            .newEvaluateDecisionCommand()
+            .decisionId(DECISION_ID)
+            .variables(Map.of("lightsaberColor", "red"))
             .send()
             .join();
-    // The Rest API returns a null future for an empty response
-    // We can verify for null, as if we'd be unauthenticated we'd get an exception
-    assertThat(response).isNull();
+
+    // then
+    assertThat(response.getDecisionOutput()).isEqualTo("\"Sith\"");
   }
 
   @Test
-  void shouldBeAuthorizedToModifyProcessWithUser() {
+  void shouldBeAuthorizedToEvaluateDecisionWithUser() {
     // given
-    final var processInstanceKey = createProcessInstance();
-    final var serviceTaskKey = getServiceTaskKey(processInstanceKey);
     final var username = UUID.randomUUID().toString();
     final var password = "password";
     authUtil.createUserWithPermissions(
         username,
         password,
         new Permissions(
-            ResourceTypeEnum.PROCESS_DEFINITION,
-            PermissionTypeEnum.UPDATE_PROCESS_INSTANCE,
-            List.of(PROCESS_ID)));
+            ResourceTypeEnum.DECISION_DEFINITION,
+            PermissionTypeEnum.CREATE_DECISION_INSTANCE,
+            List.of(DECISION_ID)));
 
     try (final var client = authUtil.createClient(username, password)) {
-      // when then
+      // when
       final var response =
           client
-              .newModifyProcessInstanceCommand(processInstanceKey)
-              .terminateElement(serviceTaskKey)
+              .newEvaluateDecisionCommand()
+              .decisionId(DECISION_ID)
+              .variables(Map.of("lightsaberColor", "red"))
               .send()
               .join();
-      // The Rest API returns a null future for an empty response
-      // We can verify for null, as if we'd be unauthenticated we'd get an exception
-      assertThat(response).isNull();
+
+      // then
+      assertThat(response.getDecisionOutput()).isEqualTo("\"Sith\"");
     }
   }
 
   @Test
-  void shouldBeUnauthorizedToModifyProcessIfNoPermissions() {
+  void shouldBeUnauthorizedToEvaluateDecisionIfNoPermissions() {
     // given
-    final var processInstanceKey = createProcessInstance();
-    final var serviceTaskKey = getServiceTaskKey(processInstanceKey);
     final var username = UUID.randomUUID().toString();
     final var password = "password";
     authUtil.createUser(username, password);
 
     try (final var client = authUtil.createClient(username, password)) {
-      // when then
+      // when
       final var response =
           client
-              .newModifyProcessInstanceCommand(processInstanceKey)
-              .terminateElement(serviceTaskKey)
+              .newEvaluateDecisionCommand()
+              .decisionId(DECISION_ID)
+              .variables(Map.of("lightsaberColor", "red"))
               .send();
 
       // then
@@ -150,27 +135,8 @@ public class ProcessInstanceModificationModifyAuthorizationIT {
           .hasMessageContaining("title: FORBIDDEN")
           .hasMessageContaining("status: 403")
           .hasMessageContaining(
-              "Insufficient permissions to perform operation 'UPDATE_PROCESS_INSTANCE' on resource 'PROCESS_DEFINITION', required resource identifiers are one of '[*, %s]'",
-              PROCESS_ID);
+              "Insufficient permissions to perform operation 'CREATE_DECISION_INSTANCE' on resource 'DECISION_DEFINITION', required resource identifiers are one of '[*, %s]'",
+              DECISION_ID);
     }
-  }
-
-  private long createProcessInstance() {
-    return defaultUserClient
-        .newCreateInstanceCommand()
-        .bpmnProcessId(PROCESS_ID)
-        .latestVersion()
-        .send()
-        .join()
-        .getProcessInstanceKey();
-  }
-
-  private static long getServiceTaskKey(final long processIstanceKey) {
-    return RecordingExporter.processInstanceRecords(ProcessInstanceIntent.ELEMENT_ACTIVATED)
-        .withProcessInstanceKey(processIstanceKey)
-        .withElementId(SERVICE_TASK_ID)
-        .limit(1)
-        .getFirst()
-        .getKey();
   }
 }
