+++ zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/authorization/permissions/IncidentResolveAuthorizationTest.java
@@ -19,11 +19,13 @@ import io.camunda.client.protocol.rest.ResourceTypeEnum;
 import io.camunda.zeebe.it.util.AuthorizationsUtil;
 import io.camunda.zeebe.it.util.AuthorizationsUtil.Permissions;
 import io.camunda.zeebe.model.bpmn.Bpmn;
+import io.camunda.zeebe.protocol.record.intent.IncidentIntent;
 import io.camunda.zeebe.qa.util.cluster.TestStandaloneBroker;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration.TestZeebe;
 import io.camunda.zeebe.test.util.junit.AutoCloseResources;
 import io.camunda.zeebe.test.util.junit.AutoCloseResources.AutoCloseResource;
+import io.camunda.zeebe.test.util.record.RecordingExporter;
 import io.camunda.zeebe.test.util.testcontainers.TestSearchContainers;
 import java.util.List;
 import java.util.UUID;
@@ -36,7 +38,7 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 @AutoCloseResources
 @Testcontainers
 @ZeebeIntegration
-public class ProcessInstanceCreateAuthorizationIT {
+public class IncidentResolveAuthorizationTest {
   @Container
   private static final ElasticsearchContainer CONTAINER =
       TestSearchContainers.createDefeaultElasticsearchContainer();
@@ -65,41 +67,33 @@ public class ProcessInstanceCreateAuthorizationIT {
     defaultUserClient
         .newDeployResourceCommand()
         .addProcessModel(
-            Bpmn.createExecutableProcess(PROCESS_ID).startEvent().endEvent().done(), "process.xml")
+            Bpmn.createExecutableProcess(PROCESS_ID)
+                .startEvent()
+                .zeebeOutputExpression("assert(foo, foo != null)", "target")
+                .endEvent()
+                .done(),
+            "process.xml")
         .send()
         .join();
   }
 
   @Test
-  void shouldBeAuthorizedToCreateInstanceWithDefaultUser() {
-    // when then
-    final var processInstanceEvent =
-        defaultUserClient
-            .newCreateInstanceCommand()
-            .bpmnProcessId(PROCESS_ID)
-            .latestVersion()
-            .send()
-            .join();
-    assertThat(processInstanceEvent.getBpmnProcessId()).isEqualTo(PROCESS_ID);
-  }
-
-  @Test
-  void shouldBeAuthorizedToCreateInstanceWithResultWithDefaultUser() {
-    // when then
-    final var processInstanceResult =
-        defaultUserClient
-            .newCreateInstanceCommand()
-            .bpmnProcessId(PROCESS_ID)
-            .latestVersion()
-            .withResult()
-            .send()
-            .join();
-    assertThat(processInstanceResult.getBpmnProcessId()).isEqualTo(PROCESS_ID);
-  }
-
-  @Test
-  void shouldBeAuthorizedToCreateInstanceWithUser() {
+  void shouldBeAuthorizedToResolveIncidentWithDefaultUser() {
     // given
+    final var incidentKey = createIncident();
+
+    // when then
+    final var response = defaultUserClient.newResolveIncidentCommand(incidentKey).send().join();
+
+    // The Rest API returns a null future for an empty response
+    // We can verify for null, as if we'd be unauthenticated we'd get an exception
+    assertThat(response).isNull();
+  }
+
+  @Test
+  void shouldBeAuthorizedToResolveIncidentWithUser() {
+    // given
+    final var incidentKey = createIncident();
     final var username = UUID.randomUUID().toString();
     final var password = "password";
     authUtil.createUserWithPermissions(
@@ -107,91 +101,55 @@ public class ProcessInstanceCreateAuthorizationIT {
         password,
         new Permissions(
             ResourceTypeEnum.PROCESS_DEFINITION,
-            PermissionTypeEnum.CREATE_PROCESS_INSTANCE,
+            PermissionTypeEnum.UPDATE_PROCESS_INSTANCE,
             List.of(PROCESS_ID)));
 
     try (final var client = authUtil.createClient(username, password)) {
-      // when then
-      final var processInstanceEvent =
-          client.newCreateInstanceCommand().bpmnProcessId(PROCESS_ID).latestVersion().send().join();
-      assertThat(processInstanceEvent.getBpmnProcessId()).isEqualTo(PROCESS_ID);
+      // when
+      final var response = client.newResolveIncidentCommand(incidentKey).send().join();
+
+      // The Rest API returns a null future for an empty response
+      // We can verify for null, as if we'd be unauthenticated we'd get an exception
+      assertThat(response).isNull();
     }
   }
 
   @Test
-  void shouldBeAuthorizedToCreateInstanceWithResultWithUser() {
+  void shouldBeUnauthorizedToResolveIncidentIfNoPermissions() {
     // given
-    final var username = UUID.randomUUID().toString();
-    final var password = "password";
-    authUtil.createUserWithPermissions(
-        username,
-        password,
-        new Permissions(
-            ResourceTypeEnum.PROCESS_DEFINITION,
-            PermissionTypeEnum.CREATE_PROCESS_INSTANCE,
-            List.of(PROCESS_ID)));
-
-    try (final var client = authUtil.createClient(username, password)) {
-      // when then
-      final var processInstanceResult =
-          client
-              .newCreateInstanceCommand()
-              .bpmnProcessId(PROCESS_ID)
-              .latestVersion()
-              .withResult()
-              .send()
-              .join();
-      assertThat(processInstanceResult.getBpmnProcessId()).isEqualTo(PROCESS_ID);
-    }
-  }
-
-  @Test
-  void shouldBeUnauthorizedToCreateInstanceIfNoPermissions() {
-    // given
+    final var incidentKey = createIncident();
     final var username = UUID.randomUUID().toString();
     final var password = "password";
     authUtil.createUser(username, password);
 
     try (final var client = authUtil.createClient(username, password)) {
+
       // when
-      final var createFuture =
-          client.newCreateInstanceCommand().bpmnProcessId(PROCESS_ID).latestVersion().send();
+      final var response = client.newResolveIncidentCommand(incidentKey).send();
 
       // then
-      assertThatThrownBy(createFuture::join)
+      assertThatThrownBy(response::join)
           .isInstanceOf(ProblemException.class)
           .hasMessageContaining("title: FORBIDDEN")
           .hasMessageContaining("status: 403")
           .hasMessageContaining(
-              "Insufficient permissions to perform operation 'CREATE_PROCESS_INSTANCE' on resource 'PROCESS_DEFINITION'");
-    }
-  }
-
-  @Test
-  void shouldBeUnauthorizedToCreateInstanceWithResultIfNoPermissions() {
-    // given
-    final var username = UUID.randomUUID().toString();
-    final var password = "password";
-    authUtil.createUser(username, password);
-
-    try (final var client = authUtil.createClient(username, password)) {
-      // when
-      final var createFuture =
-          client
-              .newCreateInstanceCommand()
-              .bpmnProcessId(PROCESS_ID)
-              .latestVersion()
-              .withResult()
-              .send();
-
-      // then
-      assertThatThrownBy(createFuture::join)
-          .isInstanceOf(ProblemException.class)
-          .hasMessageContaining("title: FORBIDDEN")
-          .hasMessageContaining("status: 403")
-          .hasMessageContaining(
-              "Insufficient permissions to perform operation 'CREATE_PROCESS_INSTANCE' on resource 'PROCESS_DEFINITION', required resource identifiers are one of '[*, %s]'",
+              "Insufficient permissions to perform operation 'UPDATE_PROCESS_INSTANCE' on resource 'PROCESS_DEFINITION', required resource identifiers are one of '[*, %s]'",
               PROCESS_ID);
     }
+  }
+
+  private long createIncident() {
+    final var processInstanceKey =
+        defaultUserClient
+            .newCreateInstanceCommand()
+            .bpmnProcessId(PROCESS_ID)
+            .latestVersion()
+            .send()
+            .join()
+            .getProcessInstanceKey();
+    return RecordingExporter.incidentRecords(IncidentIntent.CREATED)
+        .withProcessInstanceKey(processInstanceKey)
+        .getFirst()
+        .getKey();
   }
 }
