+++ zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/authorization/permissions/JobCompleteAuthorizationTest.java
@@ -19,14 +19,15 @@ import io.camunda.client.protocol.rest.ResourceTypeEnum;
 import io.camunda.zeebe.it.util.AuthorizationsUtil;
 import io.camunda.zeebe.it.util.AuthorizationsUtil.Permissions;
 import io.camunda.zeebe.model.bpmn.Bpmn;
+import io.camunda.zeebe.protocol.record.intent.JobIntent;
 import io.camunda.zeebe.qa.util.cluster.TestStandaloneBroker;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration.TestZeebe;
 import io.camunda.zeebe.test.util.junit.AutoCloseResources;
 import io.camunda.zeebe.test.util.junit.AutoCloseResources.AutoCloseResource;
+import io.camunda.zeebe.test.util.record.RecordingExporter;
 import io.camunda.zeebe.test.util.testcontainers.TestSearchContainers;
 import java.util.List;
-import java.util.Map;
 import java.util.UUID;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -37,10 +38,9 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 @AutoCloseResources
 @Testcontainers
 @ZeebeIntegration
-public class MessagePublishAuthorizationIT {
-  public static final String INTERMEDIATE_MSG_NAME = "intermediateMsg";
-  public static final String START_MSG_NAME = "startMsg";
-  public static final String CORRELATION_KEY_VARIABLE = "correlationKey";
+public class JobCompleteAuthorizationTest {
+
+  public static final String JOB_TYPE = "jobType";
 
   @Container
   private static final ElasticsearchContainer CONTAINER =
@@ -72,15 +72,8 @@ public class MessagePublishAuthorizationIT {
         .addProcessModel(
             Bpmn.createExecutableProcess(PROCESS_ID)
                 .startEvent()
-                .intermediateCatchEvent()
-                .message(
-                    m ->
-                        m.name(INTERMEDIATE_MSG_NAME)
-                            .zeebeCorrelationKeyExpression(CORRELATION_KEY_VARIABLE))
+                .serviceTask("serviceTask", t -> t.zeebeJobType(JOB_TYPE))
                 .endEvent()
-                .moveToProcess(PROCESS_ID)
-                .startEvent()
-                .message(m -> m.name(START_MSG_NAME))
                 .done(),
             "process.xml")
         .send()
@@ -88,68 +81,87 @@ public class MessagePublishAuthorizationIT {
   }
 
   @Test
-  void shouldBeAuthorizedToPublishMessageWithDefaultUser() {
+  void shouldBeAuthorizedToCompleteJobWithDefaultUser() {
     // given
-    final var correlationKey = UUID.randomUUID().toString();
-    createProcessInstance(correlationKey);
-
-    // when
-    final var response =
+    final var processInstanceKey =
         defaultUserClient
-            .newPublishMessageCommand()
-            .messageName(INTERMEDIATE_MSG_NAME)
-            .correlationKey(correlationKey)
+            .newCreateInstanceCommand()
+            .bpmnProcessId(PROCESS_ID)
+            .latestVersion()
             .send()
-            .join();
+            .join()
+            .getProcessInstanceKey();
+    final var jobKey =
+        RecordingExporter.jobRecords(JobIntent.CREATED)
+            .withProcessInstanceKey(processInstanceKey)
+            .getFirst()
+            .getKey();
 
-    // then
-    assertThat(response.getMessageKey()).isPositive();
+    // when then
+    final var response = defaultUserClient.newCompleteCommand(jobKey).send().join();
+    // The Rest API returns a null future for an empty response
+    // We can verify for null, as if we'd be unauthenticated we'd get an exception
+    assertThat(response).isNull();
   }
 
   @Test
-  void shouldBeAuthorizedToPublishMessageWithUser() {
+  void shouldBeAuthorizedToCompleteJobWithUser() {
     // given
-    final var correlationKey = UUID.randomUUID().toString();
-    createProcessInstance(correlationKey);
     final var username = UUID.randomUUID().toString();
     final var password = "password";
     authUtil.createUserWithPermissions(
         username,
         password,
-        new Permissions(ResourceTypeEnum.MESSAGE, PermissionTypeEnum.CREATE, List.of("*")));
+        new Permissions(
+            ResourceTypeEnum.PROCESS_DEFINITION,
+            PermissionTypeEnum.UPDATE_PROCESS_INSTANCE,
+            List.of(PROCESS_ID)));
+    final var processInstanceKey =
+        defaultUserClient
+            .newCreateInstanceCommand()
+            .bpmnProcessId(PROCESS_ID)
+            .latestVersion()
+            .send()
+            .join()
+            .getProcessInstanceKey();
+    final var jobKey =
+        RecordingExporter.jobRecords(JobIntent.CREATED)
+            .withProcessInstanceKey(processInstanceKey)
+            .getFirst()
+            .getKey();
 
     try (final var client = authUtil.createClient(username, password)) {
-      // when
-      final var response =
-          client
-              .newPublishMessageCommand()
-              .messageName(INTERMEDIATE_MSG_NAME)
-              .correlationKey(correlationKey)
-              .send()
-              .join();
-
-      // then
-      assertThat(response.getMessageKey()).isPositive();
+      // when then
+      final var response = client.newCompleteCommand(jobKey).send().join();
+      // The Rest API returns a null future for an empty response
+      // We can verify for null, as if we'd be unauthenticated we'd get an exception
+      assertThat(response).isNull();
     }
   }
 
   @Test
-  void shouldBeUnauthorizedToPublishMessageIfNoPermissions() {
+  void shouldBeUnauthorizedToCompleteJobIfNoPermissions() {
     // given
-    final var correlationKey = UUID.randomUUID().toString();
-    createProcessInstance(correlationKey);
     final var username = UUID.randomUUID().toString();
     final var password = "password";
     authUtil.createUser(username, password);
+    final var processInstanceKey =
+        defaultUserClient
+            .newCreateInstanceCommand()
+            .bpmnProcessId(PROCESS_ID)
+            .latestVersion()
+            .send()
+            .join()
+            .getProcessInstanceKey();
+    final var jobKey =
+        RecordingExporter.jobRecords(JobIntent.CREATED)
+            .withProcessInstanceKey(processInstanceKey)
+            .getFirst()
+            .getKey();
 
     try (final var client = authUtil.createClient(username, password)) {
       // when
-      final var response =
-          client
-              .newPublishMessageCommand()
-              .messageName(INTERMEDIATE_MSG_NAME)
-              .correlationKey(correlationKey)
-              .send();
+      final var response = client.newCompleteCommand(jobKey).send();
 
       // then
       assertThatThrownBy(response::join)
@@ -157,17 +169,8 @@ public class MessagePublishAuthorizationIT {
           .hasMessageContaining("title: FORBIDDEN")
           .hasMessageContaining("status: 403")
           .hasMessageContaining(
-              "Insufficient permissions to perform operation 'CREATE' on resource 'MESSAGE'");
+              "Insufficient permissions to perform operation 'UPDATE_PROCESS_INSTANCE' on resource 'PROCESS_DEFINITION', required resource identifiers are one of '[*, %s]'",
+              PROCESS_ID);
     }
-  }
-
-  private void createProcessInstance(final String correlationKey) {
-    defaultUserClient
-        .newCreateInstanceCommand()
-        .bpmnProcessId(PROCESS_ID)
-        .latestVersion()
-        .variables(Map.of(CORRELATION_KEY_VARIABLE, correlationKey))
-        .send()
-        .join();
   }
 }
