+++ zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/authorization/permissions/JobUpdateAuthorizationTest.java
@@ -18,11 +18,14 @@ import io.camunda.client.protocol.rest.PermissionTypeEnum;
 import io.camunda.client.protocol.rest.ResourceTypeEnum;
 import io.camunda.zeebe.it.util.AuthorizationsUtil;
 import io.camunda.zeebe.it.util.AuthorizationsUtil.Permissions;
+import io.camunda.zeebe.model.bpmn.Bpmn;
+import io.camunda.zeebe.protocol.record.intent.JobIntent;
 import io.camunda.zeebe.qa.util.cluster.TestStandaloneBroker;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration.TestZeebe;
 import io.camunda.zeebe.test.util.junit.AutoCloseResources;
 import io.camunda.zeebe.test.util.junit.AutoCloseResources.AutoCloseResource;
+import io.camunda.zeebe.test.util.record.RecordingExporter;
 import io.camunda.zeebe.test.util.testcontainers.TestSearchContainers;
 import java.util.List;
 import java.util.UUID;
@@ -35,16 +38,19 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 @AutoCloseResources
 @Testcontainers
 @ZeebeIntegration
-public class UserCreateAuthorizationIT {
+public class JobUpdateAuthorizationTest {
+
   @Container
   private static final ElasticsearchContainer CONTAINER =
       TestSearchContainers.createDefeaultElasticsearchContainer();
 
+  private static final String PROCESS_ID = "processId";
   private static AuthorizationsUtil authUtil;
   @AutoCloseResource private static CamundaClient defaultUserClient;
+  private static long jobKey;
 
   @TestZeebe(autoStart = false)
-  private final TestStandaloneBroker broker =
+  private TestStandaloneBroker broker =
       new TestStandaloneBroker()
           .withRecordingExporter(true)
           .withSecurityConfig(c -> c.getAuthorizations().setEnabled(true))
@@ -60,74 +66,75 @@ public class UserCreateAuthorizationIT {
     authUtil = new AuthorizationsUtil(broker, defaultUserClient, CONTAINER.getHttpHostAddress());
 
     authUtil.awaitUserExistsInElasticsearch(defaultUsername);
+    defaultUserClient
+        .newDeployResourceCommand()
+        .addProcessModel(
+            Bpmn.createExecutableProcess(PROCESS_ID)
+                .startEvent()
+                .serviceTask("serviceTask", t -> t.zeebeJobType("jobType"))
+                .endEvent()
+                .done(),
+            "process.xml")
+        .send()
+        .join();
+
+    final var processInstanceKey =
+        defaultUserClient
+            .newCreateInstanceCommand()
+            .bpmnProcessId(PROCESS_ID)
+            .latestVersion()
+            .send()
+            .join()
+            .getProcessInstanceKey();
+    jobKey =
+        RecordingExporter.jobRecords(JobIntent.CREATED)
+            .withProcessInstanceKey(processInstanceKey)
+            .getFirst()
+            .getKey();
   }
 
   @Test
-  void shouldBeAuthorizedToCreateUserWithDefaultUser() {
+  void shouldBeAuthorizedToUpdateJobWithDefaultUser() {
+    // when then
+    final var response =
+        defaultUserClient.newUpdateJobCommand(jobKey).updateRetries(1).send().join();
+    // The Rest API returns a null future for an empty response
+    // We can verify for null, as if we'd be unauthenticated we'd get an exception
+    assertThat(response).isNull();
+  }
+
+  @Test
+  void shouldBeAuthorizedToUpdateJobWithUser() {
     // given
     final var username = UUID.randomUUID().toString();
-
-    // when
-    final var response =
-        defaultUserClient
-            .newUserCreateCommand()
-            .username(username)
-            .name("Foo")
-            .email("bar@baz.com")
-            .password("zabraboof")
-            .send()
-            .join();
-
-    // then
-    assertThat(response.getUserKey()).isPositive();
-  }
-
-  @Test
-  void shouldBeAuthorizedToCreateUserWithPermissions() {
-    // given
-    final var authUsername = UUID.randomUUID().toString();
-    final var newUsername = UUID.randomUUID().toString();
     final var password = "password";
     authUtil.createUserWithPermissions(
-        authUsername,
+        username,
         password,
-        new Permissions(ResourceTypeEnum.USER, PermissionTypeEnum.CREATE, List.of("*")));
+        new Permissions(
+            ResourceTypeEnum.PROCESS_DEFINITION,
+            PermissionTypeEnum.UPDATE_PROCESS_INSTANCE,
+            List.of(PROCESS_ID)));
 
-    try (final var client = authUtil.createClient(authUsername, password)) {
-      // when
-      final var response =
-          client
-              .newUserCreateCommand()
-              .username(newUsername)
-              .name("Foo")
-              .email("bar@baz.com")
-              .password("zabraboof")
-              .send()
-              .join();
-
-      // then
-      assertThat(response.getUserKey()).isPositive();
+    try (final var client = authUtil.createClient(username, password)) {
+      // when then
+      final var response = client.newUpdateJobCommand(jobKey).updateRetries(1).send().join();
+      // The Rest API returns a null future for an empty response
+      // We can verify for null, as if we'd be unauthenticated we'd get an exception
+      assertThat(response).isNull();
     }
   }
 
   @Test
-  void shouldBeUnAuthorizedToCreateUserWithoutPermissions() {
+  void shouldBeUnauthorizedToUpdateJobIfNoPermissions() {
     // given
-    final var authUsername = UUID.randomUUID().toString();
-    final var newUsername = UUID.randomUUID().toString();
+    final var username = UUID.randomUUID().toString();
     final var password = "password";
-    authUtil.createUser(authUsername, password);
+    authUtil.createUser(username, password);
 
-    // when
-    try (final var client = authUtil.createClient(authUsername, password)) {
-      final var response =
-          client
-              .newUserCreateCommand()
-              .username(newUsername)
-              .name("Foo")
-              .email("bar@baz.com")
-              .password("zabraboof")
-              .send();
+    try (final var client = authUtil.createClient(username, password)) {
+      // when
+      final var response = client.newUpdateJobCommand(jobKey).updateRetries(1).send();
 
       // then
       assertThatThrownBy(response::join)
@@ -135,7 +142,8 @@ public class UserCreateAuthorizationIT {
           .hasMessageContaining("title: FORBIDDEN")
           .hasMessageContaining("status: 403")
           .hasMessageContaining(
-              "Insufficient permissions to perform operation 'CREATE' on resource 'USER'");
+              "Insufficient permissions to perform operation 'UPDATE_PROCESS_INSTANCE' on resource 'PROCESS_DEFINITION', required resource identifiers are one of '[*, %s]'",
+              PROCESS_ID);
     }
   }
 }
