+++ zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/authorization/permissions/MappingCreateAuthorizationTest.java
@@ -35,7 +35,7 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 @AutoCloseResources
 @Testcontainers
 @ZeebeIntegration
-public class MappingCreateAuthorizationIT {
+public class MappingCreateAuthorizationTest {
 
   @Container
   private static final ElasticsearchContainer CONTAINER =
