+++ zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/authorization/permissions/ProcessInstanceCancelAuthorizationTest.java
@@ -12,7 +12,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import io.camunda.application.Profile;
-import io.camunda.client.ZeebeClient;
+import io.camunda.client.CamundaClient;
 import io.camunda.client.api.command.ProblemException;
 import io.camunda.client.protocol.rest.PermissionTypeEnum;
 import io.camunda.client.protocol.rest.ResourceTypeEnum;
@@ -22,7 +22,6 @@ import io.camunda.zeebe.model.bpmn.Bpmn;
 import io.camunda.zeebe.qa.util.cluster.TestStandaloneBroker;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration.TestZeebe;
-import io.camunda.zeebe.test.util.Strings;
 import io.camunda.zeebe.test.util.junit.AutoCloseResources;
 import io.camunda.zeebe.test.util.junit.AutoCloseResources.AutoCloseResource;
 import io.camunda.zeebe.test.util.testcontainers.TestSearchContainers;
@@ -37,13 +36,14 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 @AutoCloseResources
 @Testcontainers
 @ZeebeIntegration
-final class DeploymentCreateAuthorizationIT {
+public class ProcessInstanceCancelAuthorizationTest {
   @Container
   private static final ElasticsearchContainer CONTAINER =
       TestSearchContainers.createDefeaultElasticsearchContainer();
 
+  private static final String PROCESS_ID = "processId";
   private static AuthorizationsUtil authUtil;
-  @AutoCloseResource private static ZeebeClient defaultUserClient;
+  @AutoCloseResource private static CamundaClient defaultUserClient;
 
   @TestZeebe(autoStart = false)
   private TestStandaloneBroker broker =
@@ -62,77 +62,97 @@ final class DeploymentCreateAuthorizationIT {
     authUtil = new AuthorizationsUtil(broker, defaultUserClient, CONTAINER.getHttpHostAddress());
 
     authUtil.awaitUserExistsInElasticsearch(defaultUsername);
+    defaultUserClient
+        .newDeployResourceCommand()
+        .addProcessModel(
+            Bpmn.createExecutableProcess(PROCESS_ID).startEvent().userTask().endEvent().done(),
+            "process.xml")
+        .send()
+        .join();
   }
 
   @Test
-  void shouldBeAuthorizedToDeployWithDefaultUser() {
+  void shouldBeAuthorizedToCancelInstanceWithDefaultUser() {
     // given
-    final var processId = Strings.newRandomValidBpmnId();
-
-    // when then
-    final var deploymentEvent =
+    final var processInstanceEvent =
         defaultUserClient
-            .newDeployResourceCommand()
-            .addProcessModel(
-                Bpmn.createExecutableProcess(processId).startEvent().endEvent().done(),
-                "process.bpmn")
+            .newCreateInstanceCommand()
+            .bpmnProcessId(PROCESS_ID)
+            .latestVersion()
             .send()
             .join();
-    assertThat(deploymentEvent.getProcesses().getFirst().getBpmnProcessId()).isEqualTo(processId);
+
+    // when
+    final var response =
+        defaultUserClient
+            .newCancelInstanceCommand(processInstanceEvent.getProcessInstanceKey())
+            .send()
+            .join();
+
+    // then
+    assertThat(response).isNull();
   }
 
   @Test
-  void shouldBeAuthorizedToDeployWithPermissions() {
+  void shouldBeAuthorizedToCancelInstanceWithUser() {
     // given
-    final var processId = Strings.newRandomValidBpmnId();
+    final var processInstanceEvent =
+        defaultUserClient
+            .newCreateInstanceCommand()
+            .bpmnProcessId(PROCESS_ID)
+            .latestVersion()
+            .send()
+            .join();
     final var username = UUID.randomUUID().toString();
     final var password = "password";
     authUtil.createUserWithPermissions(
         username,
         password,
-        new Permissions(ResourceTypeEnum.DEPLOYMENT, PermissionTypeEnum.CREATE, List.of("*")));
+        new Permissions(
+            ResourceTypeEnum.PROCESS_DEFINITION,
+            PermissionTypeEnum.UPDATE_PROCESS_INSTANCE,
+            List.of(PROCESS_ID)));
 
     try (final var client = authUtil.createClient(username, password)) {
       // when
-      final var deploymentEvent =
+      final var response =
           client
-              .newDeployResourceCommand()
-              .addProcessModel(
-                  Bpmn.createExecutableProcess(processId).startEvent().endEvent().done(),
-                  "process.bpmn")
+              .newCancelInstanceCommand(processInstanceEvent.getProcessInstanceKey())
               .send()
               .join();
 
       // then
-      assertThat(deploymentEvent.getProcesses().getFirst().getBpmnProcessId()).isEqualTo(processId);
+      assertThat(response).isNull();
     }
   }
 
   @Test
-  void shouldBeUnAuthorizedToDeployWithPermissions() {
+  void shouldBeUnauthorizedToCancelInstanceIfNoPermissions() {
     // given
-    final var processId = Strings.newRandomValidBpmnId();
+    final var processInstanceEvent =
+        defaultUserClient
+            .newCreateInstanceCommand()
+            .bpmnProcessId(PROCESS_ID)
+            .latestVersion()
+            .send()
+            .join();
     final var username = UUID.randomUUID().toString();
     final var password = "password";
     authUtil.createUser(username, password);
 
-    // when
     try (final var client = authUtil.createClient(username, password)) {
-      final var deployFuture =
-          client
-              .newDeployResourceCommand()
-              .addProcessModel(
-                  Bpmn.createExecutableProcess(processId).startEvent().endEvent().done(),
-                  "process.bpmn")
-              .send();
+      // when
+      final var response =
+          client.newCancelInstanceCommand(processInstanceEvent.getProcessInstanceKey()).send();
 
       // then
-      assertThatThrownBy(deployFuture::join)
+      assertThatThrownBy(response::join)
           .isInstanceOf(ProblemException.class)
           .hasMessageContaining("title: FORBIDDEN")
           .hasMessageContaining("status: 403")
           .hasMessageContaining(
-              "Insufficient permissions to perform operation 'CREATE' on resource 'DEPLOYMENT'");
+              "Insufficient permissions to perform operation 'UPDATE_PROCESS_INSTANCE' on resource 'PROCESS_DEFINITION', required resource identifiers are one of '[*, %s]'",
+              PROCESS_ID);
     }
   }
 }
