+++ zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/authorization/permissions/ProcessInstanceCreateAuthorizationTest.java
@@ -19,13 +19,11 @@ import io.camunda.client.protocol.rest.ResourceTypeEnum;
 import io.camunda.zeebe.it.util.AuthorizationsUtil;
 import io.camunda.zeebe.it.util.AuthorizationsUtil.Permissions;
 import io.camunda.zeebe.model.bpmn.Bpmn;
-import io.camunda.zeebe.protocol.record.intent.JobIntent;
 import io.camunda.zeebe.qa.util.cluster.TestStandaloneBroker;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration.TestZeebe;
 import io.camunda.zeebe.test.util.junit.AutoCloseResources;
 import io.camunda.zeebe.test.util.junit.AutoCloseResources.AutoCloseResource;
-import io.camunda.zeebe.test.util.record.RecordingExporter;
 import io.camunda.zeebe.test.util.testcontainers.TestSearchContainers;
 import java.util.List;
 import java.util.UUID;
@@ -38,8 +36,7 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 @AutoCloseResources
 @Testcontainers
 @ZeebeIntegration
-public class JobFailAuthorizationIT {
-
+public class ProcessInstanceCreateAuthorizationTest {
   @Container
   private static final ElasticsearchContainer CONTAINER =
       TestSearchContainers.createDefeaultElasticsearchContainer();
@@ -47,7 +44,6 @@ public class JobFailAuthorizationIT {
   private static final String PROCESS_ID = "processId";
   private static AuthorizationsUtil authUtil;
   @AutoCloseResource private static CamundaClient defaultUserClient;
-  private static long jobKey;
 
   @TestZeebe(autoStart = false)
   private TestStandaloneBroker broker =
@@ -69,41 +65,40 @@ public class JobFailAuthorizationIT {
     defaultUserClient
         .newDeployResourceCommand()
         .addProcessModel(
-            Bpmn.createExecutableProcess(PROCESS_ID)
-                .startEvent()
-                .serviceTask("serviceTask", t -> t.zeebeJobType("jobType"))
-                .endEvent()
-                .done(),
-            "process.xml")
+            Bpmn.createExecutableProcess(PROCESS_ID).startEvent().endEvent().done(), "process.xml")
         .send()
         .join();
+  }
 
-    final var processInstanceKey =
+  @Test
+  void shouldBeAuthorizedToCreateInstanceWithDefaultUser() {
+    // when then
+    final var processInstanceEvent =
         defaultUserClient
             .newCreateInstanceCommand()
             .bpmnProcessId(PROCESS_ID)
             .latestVersion()
             .send()
-            .join()
-            .getProcessInstanceKey();
-    jobKey =
-        RecordingExporter.jobRecords(JobIntent.CREATED)
-            .withProcessInstanceKey(processInstanceKey)
-            .getFirst()
-            .getKey();
+            .join();
+    assertThat(processInstanceEvent.getBpmnProcessId()).isEqualTo(PROCESS_ID);
   }
 
   @Test
-  void shouldBeAuthorizedToFailJobWithDefaultUser() {
+  void shouldBeAuthorizedToCreateInstanceWithResultWithDefaultUser() {
     // when then
-    final var response = defaultUserClient.newFailCommand(jobKey).retries(1).send().join();
-    // The Rest API returns a null future for an empty response
-    // We can verify for null, as if we'd be unauthenticated we'd get an exception
-    assertThat(response).isNull();
+    final var processInstanceResult =
+        defaultUserClient
+            .newCreateInstanceCommand()
+            .bpmnProcessId(PROCESS_ID)
+            .latestVersion()
+            .withResult()
+            .send()
+            .join();
+    assertThat(processInstanceResult.getBpmnProcessId()).isEqualTo(PROCESS_ID);
   }
 
   @Test
-  void shouldBeAuthorizedToFailJobWithUser() {
+  void shouldBeAuthorizedToCreateInstanceWithUser() {
     // given
     final var username = UUID.randomUUID().toString();
     final var password = "password";
@@ -112,20 +107,46 @@ public class JobFailAuthorizationIT {
         password,
         new Permissions(
             ResourceTypeEnum.PROCESS_DEFINITION,
-            PermissionTypeEnum.UPDATE_PROCESS_INSTANCE,
+            PermissionTypeEnum.CREATE_PROCESS_INSTANCE,
             List.of(PROCESS_ID)));
 
     try (final var client = authUtil.createClient(username, password)) {
       // when then
-      final var response = client.newFailCommand(jobKey).retries(1).send().join();
-      // The Rest API returns a null future for an empty response
-      // We can verify for null, as if we'd be unauthenticated we'd get an exception
-      assertThat(response).isNull();
+      final var processInstanceEvent =
+          client.newCreateInstanceCommand().bpmnProcessId(PROCESS_ID).latestVersion().send().join();
+      assertThat(processInstanceEvent.getBpmnProcessId()).isEqualTo(PROCESS_ID);
     }
   }
 
   @Test
-  void shouldBeUnauthorizedToFailJobIfNoPermissions() {
+  void shouldBeAuthorizedToCreateInstanceWithResultWithUser() {
+    // given
+    final var username = UUID.randomUUID().toString();
+    final var password = "password";
+    authUtil.createUserWithPermissions(
+        username,
+        password,
+        new Permissions(
+            ResourceTypeEnum.PROCESS_DEFINITION,
+            PermissionTypeEnum.CREATE_PROCESS_INSTANCE,
+            List.of(PROCESS_ID)));
+
+    try (final var client = authUtil.createClient(username, password)) {
+      // when then
+      final var processInstanceResult =
+          client
+              .newCreateInstanceCommand()
+              .bpmnProcessId(PROCESS_ID)
+              .latestVersion()
+              .withResult()
+              .send()
+              .join();
+      assertThat(processInstanceResult.getBpmnProcessId()).isEqualTo(PROCESS_ID);
+    }
+  }
+
+  @Test
+  void shouldBeUnauthorizedToCreateInstanceIfNoPermissions() {
     // given
     final var username = UUID.randomUUID().toString();
     final var password = "password";
@@ -133,15 +154,43 @@ public class JobFailAuthorizationIT {
 
     try (final var client = authUtil.createClient(username, password)) {
       // when
-      final var response = client.newFailCommand(jobKey).retries(1).send();
+      final var createFuture =
+          client.newCreateInstanceCommand().bpmnProcessId(PROCESS_ID).latestVersion().send();
 
       // then
-      assertThatThrownBy(response::join)
+      assertThatThrownBy(createFuture::join)
           .isInstanceOf(ProblemException.class)
           .hasMessageContaining("title: FORBIDDEN")
           .hasMessageContaining("status: 403")
           .hasMessageContaining(
-              "Insufficient permissions to perform operation 'UPDATE_PROCESS_INSTANCE' on resource 'PROCESS_DEFINITION', required resource identifiers are one of '[*, %s]'",
+              "Insufficient permissions to perform operation 'CREATE_PROCESS_INSTANCE' on resource 'PROCESS_DEFINITION'");
+    }
+  }
+
+  @Test
+  void shouldBeUnauthorizedToCreateInstanceWithResultIfNoPermissions() {
+    // given
+    final var username = UUID.randomUUID().toString();
+    final var password = "password";
+    authUtil.createUser(username, password);
+
+    try (final var client = authUtil.createClient(username, password)) {
+      // when
+      final var createFuture =
+          client
+              .newCreateInstanceCommand()
+              .bpmnProcessId(PROCESS_ID)
+              .latestVersion()
+              .withResult()
+              .send();
+
+      // then
+      assertThatThrownBy(createFuture::join)
+          .isInstanceOf(ProblemException.class)
+          .hasMessageContaining("title: FORBIDDEN")
+          .hasMessageContaining("status: 403")
+          .hasMessageContaining(
+              "Insufficient permissions to perform operation 'CREATE_PROCESS_INSTANCE' on resource 'PROCESS_DEFINITION', required resource identifiers are one of '[*, %s]'",
               PROCESS_ID);
     }
   }
