+++ zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/authorization/permissions/ProcessInstanceModificationModifyAuthorizationTest.java
@@ -19,11 +19,13 @@ import io.camunda.client.protocol.rest.ResourceTypeEnum;
 import io.camunda.zeebe.it.util.AuthorizationsUtil;
 import io.camunda.zeebe.it.util.AuthorizationsUtil.Permissions;
 import io.camunda.zeebe.model.bpmn.Bpmn;
+import io.camunda.zeebe.protocol.record.intent.ProcessInstanceIntent;
 import io.camunda.zeebe.qa.util.cluster.TestStandaloneBroker;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration.TestZeebe;
 import io.camunda.zeebe.test.util.junit.AutoCloseResources;
 import io.camunda.zeebe.test.util.junit.AutoCloseResources.AutoCloseResource;
+import io.camunda.zeebe.test.util.record.RecordingExporter;
 import io.camunda.zeebe.test.util.testcontainers.TestSearchContainers;
 import java.util.List;
 import java.util.UUID;
@@ -36,14 +38,17 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 @AutoCloseResources
 @Testcontainers
 @ZeebeIntegration
-public class AuthorizationRemovePermissionAuthorizationIT {
+public class ProcessInstanceModificationModifyAuthorizationTest {
+  public static final String JOB_TYPE = "jobType";
+  public static final String SERVICE_TASK_ID = "serviceTask";
+
   @Container
   private static final ElasticsearchContainer CONTAINER =
       TestSearchContainers.createDefeaultElasticsearchContainer();
 
   private static final String PROCESS_ID = "processId";
-  @AutoCloseResource private static CamundaClient defaultUserClient;
   private static AuthorizationsUtil authUtil;
+  @AutoCloseResource private static CamundaClient defaultUserClient;
 
   @TestZeebe(autoStart = false)
   private TestStandaloneBroker broker =
@@ -65,59 +70,55 @@ public class AuthorizationRemovePermissionAuthorizationIT {
     defaultUserClient
         .newDeployResourceCommand()
         .addProcessModel(
-            Bpmn.createExecutableProcess(PROCESS_ID).startEvent().endEvent().done(), "process.xml")
+            Bpmn.createExecutableProcess(PROCESS_ID)
+                .startEvent()
+                .serviceTask(SERVICE_TASK_ID, t -> t.zeebeJobType(JOB_TYPE))
+                .endEvent()
+                .done(),
+            "process.xml")
         .send()
         .join();
   }
 
   @Test
-  void shouldBeAuthorizedToRemovePermissionsWithDefaultUser() {
+  void shouldBeAuthorizedToModifyProcessWithDefaultUser() {
     // given
-    final var username = UUID.randomUUID().toString();
-    final var resourceType = ResourceTypeEnum.DEPLOYMENT;
-    final var permissionType = PermissionTypeEnum.DELETE;
-    final var resourceId = "resourceId";
-    final var userKey =
-        authUtil.createUserWithPermissions(
-            username,
-            "password",
-            new Permissions(resourceType, permissionType, List.of(resourceId)));
+    final var processInstanceKey = createProcessInstance();
+    final var serviceTaskKey = getServiceTaskKey(processInstanceKey);
 
     // when then
     final var response =
         defaultUserClient
-            .newRemovePermissionsCommand(userKey)
-            .resourceType(resourceType)
-            .permission(permissionType)
-            .resourceId(resourceId)
+            .newModifyProcessInstanceCommand(processInstanceKey)
+            .terminateElement(serviceTaskKey)
             .send()
             .join();
-
     // The Rest API returns a null future for an empty response
     // We can verify for null, as if we'd be unauthenticated we'd get an exception
     assertThat(response).isNull();
   }
 
   @Test
-  void shouldBeAuthorizedToRemovePermissionsWithUser() {
+  void shouldBeAuthorizedToModifyProcessWithUser() {
     // given
+    final var processInstanceKey = createProcessInstance();
+    final var serviceTaskKey = getServiceTaskKey(processInstanceKey);
     final var username = UUID.randomUUID().toString();
     final var password = "password";
-    final var resourceType = ResourceTypeEnum.AUTHORIZATION;
-    final var permissionType = PermissionTypeEnum.UPDATE;
-    final var resourceId = "*";
-    final var userKey =
-        authUtil.createUserWithPermissions(
-            username, password, new Permissions(resourceType, permissionType, List.of(resourceId)));
+    authUtil.createUserWithPermissions(
+        username,
+        password,
+        new Permissions(
+            ResourceTypeEnum.PROCESS_DEFINITION,
+            PermissionTypeEnum.UPDATE_PROCESS_INSTANCE,
+            List.of(PROCESS_ID)));
 
     try (final var client = authUtil.createClient(username, password)) {
       // when then
       final var response =
           client
-              .newRemovePermissionsCommand(userKey)
-              .resourceType(resourceType)
-              .permission(permissionType)
-              .resourceId(resourceId)
+              .newModifyProcessInstanceCommand(processInstanceKey)
+              .terminateElement(serviceTaskKey)
               .send()
               .join();
       // The Rest API returns a null future for an empty response
@@ -127,21 +128,20 @@ public class AuthorizationRemovePermissionAuthorizationIT {
   }
 
   @Test
-  void shouldBeUnauthorizedToRemovePermissionsIfNoPermissions() {
+  void shouldBeUnauthorizedToModifyProcessIfNoPermissions() {
     // given
+    final var processInstanceKey = createProcessInstance();
+    final var serviceTaskKey = getServiceTaskKey(processInstanceKey);
     final var username = UUID.randomUUID().toString();
     final var password = "password";
     authUtil.createUser(username, password);
 
-    // when
     try (final var client = authUtil.createClient(username, password)) {
+      // when then
       final var response =
           client
-              // We can use any owner key. The authorization check happens before we use it.
-              .newRemovePermissionsCommand(1L)
-              .resourceType(ResourceTypeEnum.DEPLOYMENT)
-              .permission(PermissionTypeEnum.CREATE)
-              .resourceId("*")
+              .newModifyProcessInstanceCommand(processInstanceKey)
+              .terminateElement(serviceTaskKey)
               .send();
 
       // then
@@ -150,7 +150,27 @@ public class AuthorizationRemovePermissionAuthorizationIT {
           .hasMessageContaining("title: FORBIDDEN")
           .hasMessageContaining("status: 403")
           .hasMessageContaining(
-              "Insufficient permissions to perform operation 'UPDATE' on resource 'AUTHORIZATION'");
+              "Insufficient permissions to perform operation 'UPDATE_PROCESS_INSTANCE' on resource 'PROCESS_DEFINITION', required resource identifiers are one of '[*, %s]'",
+              PROCESS_ID);
     }
+  }
+
+  private long createProcessInstance() {
+    return defaultUserClient
+        .newCreateInstanceCommand()
+        .bpmnProcessId(PROCESS_ID)
+        .latestVersion()
+        .send()
+        .join()
+        .getProcessInstanceKey();
+  }
+
+  private static long getServiceTaskKey(final long processIstanceKey) {
+    return RecordingExporter.processInstanceRecords(ProcessInstanceIntent.ELEMENT_ACTIVATED)
+        .withProcessInstanceKey(processIstanceKey)
+        .withElementId(SERVICE_TASK_ID)
+        .limit(1)
+        .getFirst()
+        .getKey();
   }
 }
