+++ zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/authorization/permissions/SignalBroadcastAuthorizationTest.java
@@ -19,14 +19,18 @@ import io.camunda.client.protocol.rest.ResourceTypeEnum;
 import io.camunda.zeebe.it.util.AuthorizationsUtil;
 import io.camunda.zeebe.it.util.AuthorizationsUtil.Permissions;
 import io.camunda.zeebe.model.bpmn.Bpmn;
+import io.camunda.zeebe.protocol.record.RejectionType;
+import io.camunda.zeebe.protocol.record.intent.ProcessInstanceIntent;
+import io.camunda.zeebe.protocol.record.value.BpmnElementType;
+import io.camunda.zeebe.protocol.record.value.BpmnEventType;
 import io.camunda.zeebe.qa.util.cluster.TestStandaloneBroker;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration.TestZeebe;
 import io.camunda.zeebe.test.util.junit.AutoCloseResources;
 import io.camunda.zeebe.test.util.junit.AutoCloseResources.AutoCloseResource;
+import io.camunda.zeebe.test.util.record.RecordingExporter;
 import io.camunda.zeebe.test.util.testcontainers.TestSearchContainers;
 import java.util.List;
-import java.util.Map;
 import java.util.UUID;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -37,7 +41,9 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 @AutoCloseResources
 @Testcontainers
 @ZeebeIntegration
-public class VariableDocumentUpdateAuthorizationIT {
+public class SignalBroadcastAuthorizationTest {
+  public static final String SIGNAL_NAME = "signal";
+
   @Container
   private static final ElasticsearchContainer CONTAINER =
       TestSearchContainers.createDefeaultElasticsearchContainer();
@@ -68,8 +74,12 @@ public class VariableDocumentUpdateAuthorizationIT {
         .addProcessModel(
             Bpmn.createExecutableProcess(PROCESS_ID)
                 .startEvent()
-                .userTask("task")
+                .intermediateCatchEvent()
+                .signal(s -> s.name(SIGNAL_NAME))
                 .endEvent()
+                .moveToProcess(PROCESS_ID)
+                .startEvent()
+                .signal(s -> s.name(SIGNAL_NAME))
                 .done(),
             "process.xml")
         .send()
@@ -77,27 +87,22 @@ public class VariableDocumentUpdateAuthorizationIT {
   }
 
   @Test
-  void shouldBeAuthorizedToUpdateVariablesWithDefaultUser() {
+  void shouldBeAuthorizedToBroadcastSignalWithDefaultUser() {
     // given
-    final var processInstanceKey = createProcessInstance();
+    createProcessInstance();
 
-    // when then
+    // when
     final var response =
-        defaultUserClient
-            .newSetVariablesCommand(processInstanceKey)
-            .variables(Map.of("foo", "bar"))
-            .send()
-            .join();
+        defaultUserClient.newBroadcastSignalCommand().signalName(SIGNAL_NAME).send().join();
 
-    // The Rest API returns a null future for an empty response
-    // We can verify for null, as if we'd be unauthenticated we'd get an exception
-    assertThat(response).isNull();
+    // then
+    assertThat(response.getKey()).isPositive();
   }
 
   @Test
-  void shouldBeAuthorizedToUpdateVariablesWithUser() {
+  void shouldBeAuthorizedToBroadcastSignalWithUser() {
     // given
-    final var processInstanceKey = createProcessInstance();
+    createProcessInstance();
     final var username = UUID.randomUUID().toString();
     final var password = "password";
     authUtil.createUserWithPermissions(
@@ -106,36 +111,62 @@ public class VariableDocumentUpdateAuthorizationIT {
         new Permissions(
             ResourceTypeEnum.PROCESS_DEFINITION,
             PermissionTypeEnum.UPDATE_PROCESS_INSTANCE,
+            List.of(PROCESS_ID)),
+        new Permissions(
+            ResourceTypeEnum.PROCESS_DEFINITION,
+            PermissionTypeEnum.CREATE_PROCESS_INSTANCE,
             List.of(PROCESS_ID)));
 
     try (final var client = authUtil.createClient(username, password)) {
-      // when then
-      final var response =
-          client
-              .newSetVariablesCommand(processInstanceKey)
-              .variables(Map.of("foo", "bar"))
-              .send()
-              .join();
+      // when
+      final var response = client.newBroadcastSignalCommand().signalName(SIGNAL_NAME).send().join();
 
-      // The Rest API returns a null future for an empty response
-      // We can verify for null, as if we'd be unauthenticated we'd get an exception
-      assertThat(response).isNull();
+      // then
+      assertThat(response.getKey()).isPositive();
     }
   }
 
   @Test
-  void shouldBeUnauthorizedToUpdateVariablesIfNoPermissions() {
+  void shouldBeUnauthorizedToBroadcastSignalIfNoPermissions() {
     // given
-    final var processInstanceKey = createProcessInstance();
+    createProcessInstance();
     final var username = UUID.randomUUID().toString();
     final var password = "password";
     authUtil.createUser(username, password);
 
-    // when
+    try (final var client = authUtil.createClient(username, password)) {
+
+      // when
+      final var response = client.newBroadcastSignalCommand().signalName(SIGNAL_NAME).send();
+
+      // then
+      assertThatThrownBy(response::join)
+          .isInstanceOf(ProblemException.class)
+          .hasMessageContaining("title: FORBIDDEN")
+          .hasMessageContaining("status: 403")
+          .hasMessageContaining(
+              "Insufficient permissions to perform operation 'CREATE_PROCESS_INSTANCE' on resource 'PROCESS_DEFINITION', required resource identifiers are one of '[*, %s]'",
+              PROCESS_ID);
+    }
+  }
+
+  @Test
+  void shouldNotBroadcastSignalIfUnauthorizedForOne() {
+    // given
+    final var username = UUID.randomUUID().toString();
+    final var password = "password";
+    authUtil.createUserWithPermissions(
+        username,
+        password,
+        new Permissions(
+            ResourceTypeEnum.PROCESS_DEFINITION,
+            PermissionTypeEnum.CREATE_PROCESS_INSTANCE,
+            List.of(PROCESS_ID)));
+    final var processInstanceKey = createProcessInstance();
+
     try (final var client = authUtil.createClient(username, password)) {
       // when
-      final var response =
-          client.newSetVariablesCommand(processInstanceKey).variables(Map.of("foo", "bar")).send();
+      final var response = client.newBroadcastSignalCommand().signalName(SIGNAL_NAME).send();
 
       // then
       assertThatThrownBy(response::join)
@@ -145,16 +176,36 @@ public class VariableDocumentUpdateAuthorizationIT {
           .hasMessageContaining(
               "Insufficient permissions to perform operation 'UPDATE_PROCESS_INSTANCE' on resource 'PROCESS_DEFINITION', required resource identifiers are one of '[*, %s]'",
               PROCESS_ID);
+
+      assertThat(
+              RecordingExporter.records()
+                  .limit(r -> r.getRejectionType() == RejectionType.UNAUTHORIZED)
+                  .processInstanceRecords()
+                  .withProcessInstanceKey(processInstanceKey)
+                  .withElementType(BpmnElementType.INTERMEDIATE_CATCH_EVENT)
+                  .withEventType(BpmnEventType.SIGNAL)
+                  .withIntent(ProcessInstanceIntent.ELEMENT_COMPLETED)
+                  .exists())
+          .isFalse();
     }
   }
 
-  private long createProcessInstance() {
-    return defaultUserClient
-        .newCreateInstanceCommand()
-        .bpmnProcessId(PROCESS_ID)
-        .latestVersion()
-        .send()
-        .join()
-        .getProcessInstanceKey();
+  private Long createProcessInstance() {
+    final var processInstanceKey =
+        defaultUserClient
+            .newCreateInstanceCommand()
+            .bpmnProcessId(PROCESS_ID)
+            .latestVersion()
+            .send()
+            .join()
+            .getProcessInstanceKey();
+
+    RecordingExporter.processInstanceRecords(ProcessInstanceIntent.ELEMENT_ACTIVATED)
+        .withProcessInstanceKey(processInstanceKey)
+        .withElementType(BpmnElementType.INTERMEDIATE_CATCH_EVENT)
+        .withEventType(BpmnEventType.SIGNAL)
+        .await();
+
+    return processInstanceKey;
   }
 }
