+++ zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/authorization/permissions/UserTaskAssignAuthorizationTest.java
@@ -19,11 +19,13 @@ import io.camunda.client.protocol.rest.ResourceTypeEnum;
 import io.camunda.zeebe.it.util.AuthorizationsUtil;
 import io.camunda.zeebe.it.util.AuthorizationsUtil.Permissions;
 import io.camunda.zeebe.model.bpmn.Bpmn;
+import io.camunda.zeebe.protocol.record.intent.UserTaskIntent;
 import io.camunda.zeebe.qa.util.cluster.TestStandaloneBroker;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration.TestZeebe;
 import io.camunda.zeebe.test.util.junit.AutoCloseResources;
 import io.camunda.zeebe.test.util.junit.AutoCloseResources.AutoCloseResource;
+import io.camunda.zeebe.test.util.record.RecordingExporter;
 import io.camunda.zeebe.test.util.testcontainers.TestSearchContainers;
 import java.util.List;
 import java.util.UUID;
@@ -36,14 +38,16 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 @AutoCloseResources
 @Testcontainers
 @ZeebeIntegration
-public class AuthorizationAddPermissionAuthorizationIT {
+public class UserTaskAssignAuthorizationTest {
+  public static final String USER_TASK_ID = "userTask";
+
   @Container
   private static final ElasticsearchContainer CONTAINER =
       TestSearchContainers.createDefeaultElasticsearchContainer();
 
   private static final String PROCESS_ID = "processId";
-  @AutoCloseResource private static CamundaClient defaultUserClient;
   private static AuthorizationsUtil authUtil;
+  @AutoCloseResource private static CamundaClient defaultUserClient;
 
   @TestZeebe(autoStart = false)
   private TestStandaloneBroker broker =
@@ -65,75 +69,84 @@ public class AuthorizationAddPermissionAuthorizationIT {
     defaultUserClient
         .newDeployResourceCommand()
         .addProcessModel(
-            Bpmn.createExecutableProcess(PROCESS_ID).startEvent().endEvent().done(), "process.xml")
+            Bpmn.createExecutableProcess(PROCESS_ID)
+                .startEvent()
+                .userTask(USER_TASK_ID)
+                .zeebeUserTask()
+                .endEvent()
+                .done(),
+            "process.xml")
         .send()
         .join();
   }
 
   @Test
-  void shouldBeAuthorizedToAddPermissionsWithDefaultUser() {
+  void shouldBeAuthorizedToAssignUserTaskWithDefaultUser() {
     // given
-    final var username = UUID.randomUUID().toString();
-    final var userKey = authUtil.createUser(username, "password");
+    final var processInstanceKey = createProcessInstance();
+    final var userTaskKey = getUserTaskKey(processInstanceKey);
 
     // when then
-    final var addPermissionsResponse =
+    final var response =
         defaultUserClient
-            .newAddPermissionsCommand(userKey)
-            .resourceType(ResourceTypeEnum.DEPLOYMENT)
-            .permission(PermissionTypeEnum.DELETE)
-            .resourceId("*")
+            .newUserTaskAssignCommand(userTaskKey)
+            .assignee("assignee")
+            .allowOverride(true)
             .send()
             .join();
+
     // The Rest API returns a null future for an empty response
     // We can verify for null, as if we'd be unauthenticated we'd get an exception
-    assertThat(addPermissionsResponse).isNull();
+    assertThat(response).isNull();
   }
 
   @Test
-  void shouldBeAuthorizedToAddPermissionsWithUser() {
+  void shouldBeAuthorizedToAssignUserTaskWithUser() {
     // given
+    final var processInstanceKey = createProcessInstance();
+    final var userTaskKey = getUserTaskKey(processInstanceKey);
     final var username = UUID.randomUUID().toString();
     final var password = "password";
-    final var userKey =
-        authUtil.createUserWithPermissions(
-            username,
-            password,
-            new Permissions(
-                ResourceTypeEnum.AUTHORIZATION, PermissionTypeEnum.UPDATE, List.of("*")));
+    authUtil.createUserWithPermissions(
+        username,
+        password,
+        new Permissions(
+            ResourceTypeEnum.PROCESS_DEFINITION,
+            PermissionTypeEnum.UPDATE_USER_TASK,
+            List.of(PROCESS_ID)));
 
     try (final var client = authUtil.createClient(username, password)) {
       // when then
-      final var addPermissionsResponse =
+      final var response =
           client
-              .newAddPermissionsCommand(userKey)
-              .resourceType(ResourceTypeEnum.DEPLOYMENT)
-              .permission(PermissionTypeEnum.CREATE)
-              .resourceId("*")
+              .newUserTaskAssignCommand(userTaskKey)
+              .assignee("assignee")
+              .allowOverride(true)
               .send()
               .join();
+
       // The Rest API returns a null future for an empty response
       // We can verify for null, as if we'd be unauthenticated we'd get an exception
-      assertThat(addPermissionsResponse).isNull();
+      assertThat(response).isNull();
     }
   }
 
   @Test
-  void shouldBeUnauthorizedToAddPermissionsIfNoPermissions() {
+  void shouldBeUnauthorizedToAssignUserTaskIfNoPermissions() {
     // given
+    final var processInstanceKey = createProcessInstance();
+    final var userTaskKey = getUserTaskKey(processInstanceKey);
     final var username = UUID.randomUUID().toString();
     final var password = "password";
     authUtil.createUser(username, password);
 
-    // when
     try (final var client = authUtil.createClient(username, password)) {
+      // when we use the unauthorized client
       final var response =
           client
-              // We can use any owner key. The authorization check happens before we use it.
-              .newAddPermissionsCommand(1L)
-              .resourceType(ResourceTypeEnum.DEPLOYMENT)
-              .permission(PermissionTypeEnum.CREATE)
-              .resourceId("*")
+              .newUserTaskAssignCommand(userTaskKey)
+              .assignee("assignee")
+              .allowOverride(true)
               .send();
 
       // then
@@ -142,7 +155,28 @@ public class AuthorizationAddPermissionAuthorizationIT {
           .hasMessageContaining("title: FORBIDDEN")
           .hasMessageContaining("status: 403")
           .hasMessageContaining(
-              "Insufficient permissions to perform operation 'UPDATE' on resource 'AUTHORIZATION'");
+              "Insufficient permissions to perform operation 'UPDATE_USER_TASK' on resource 'PROCESS_DEFINITION', required resource identifiers are one of '[*, %s]'",
+              PROCESS_ID);
     }
+  }
+
+  private long createProcessInstance() {
+    return defaultUserClient
+        .newCreateInstanceCommand()
+        .bpmnProcessId(PROCESS_ID)
+        .latestVersion()
+        .send()
+        .join()
+        .getProcessInstanceKey();
+  }
+
+  private static long getUserTaskKey(final long processInstanceKey) {
+    return RecordingExporter.userTaskRecords(UserTaskIntent.CREATED)
+        .withProcessInstanceKey(processInstanceKey)
+        .withElementId(USER_TASK_ID)
+        .limit(1)
+        .findFirst()
+        .orElseThrow()
+        .getKey();
   }
 }
