+++ zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/authorization/permissions/UserTaskUpdateAuthorizationTest.java
@@ -19,11 +19,13 @@ import io.camunda.client.protocol.rest.ResourceTypeEnum;
 import io.camunda.zeebe.it.util.AuthorizationsUtil;
 import io.camunda.zeebe.it.util.AuthorizationsUtil.Permissions;
 import io.camunda.zeebe.model.bpmn.Bpmn;
+import io.camunda.zeebe.protocol.record.intent.UserTaskIntent;
 import io.camunda.zeebe.qa.util.cluster.TestStandaloneBroker;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration.TestZeebe;
 import io.camunda.zeebe.test.util.junit.AutoCloseResources;
 import io.camunda.zeebe.test.util.junit.AutoCloseResources.AutoCloseResource;
+import io.camunda.zeebe.test.util.record.RecordingExporter;
 import io.camunda.zeebe.test.util.testcontainers.TestSearchContainers;
 import java.util.List;
 import java.util.UUID;
@@ -36,23 +38,19 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 @AutoCloseResources
 @Testcontainers
 @ZeebeIntegration
-public class ProcessInstanceMigrationMigrateAuthorizationIT {
-  public static final String JOB_TYPE = "jobType";
-  public static final String SOURCE_TASK = "sourceTask";
-  public static final String TARGET_TASK = "targetTask";
+public class UserTaskUpdateAuthorizationTest {
+  public static final String USER_TASK_ID = "userTask";
 
   @Container
   private static final ElasticsearchContainer CONTAINER =
       TestSearchContainers.createDefeaultElasticsearchContainer();
 
   private static final String PROCESS_ID = "processId";
-  private static final String TARGET_PROCESS_ID = "targetProcessId";
   private static AuthorizationsUtil authUtil;
   @AutoCloseResource private static CamundaClient defaultUserClient;
-  private static long targetProcDefKey;
 
   @TestZeebe(autoStart = false)
-  private final TestStandaloneBroker broker =
+  private TestStandaloneBroker broker =
       new TestStandaloneBroker()
           .withRecordingExporter(true)
           .withSecurityConfig(c -> c.getAuthorizations().setEnabled(true))
@@ -68,54 +66,29 @@ public class ProcessInstanceMigrationMigrateAuthorizationIT {
     authUtil = new AuthorizationsUtil(broker, defaultUserClient, CONTAINER.getHttpHostAddress());
 
     authUtil.awaitUserExistsInElasticsearch(defaultUsername);
-    final var deploymentEvent =
-        defaultUserClient
-            .newDeployResourceCommand()
-            .addProcessModel(
-                Bpmn.createExecutableProcess(PROCESS_ID)
-                    .startEvent()
-                    .serviceTask(SOURCE_TASK, t -> t.zeebeJobType(JOB_TYPE))
-                    .endEvent()
-                    .done(),
-                "process.xml")
-            .addProcessModel(
-                Bpmn.createExecutableProcess(TARGET_PROCESS_ID)
-                    .startEvent()
-                    .serviceTask(TARGET_TASK, t -> t.zeebeJobType(JOB_TYPE))
-                    .endEvent()
-                    .done(),
-                "targetProcess.xml")
-            .send()
-            .join();
-    targetProcDefKey =
-        deploymentEvent.getProcesses().stream()
-            .filter(process -> process.getBpmnProcessId().equals(TARGET_PROCESS_ID))
-            .findFirst()
-            .orElseThrow()
-            .getProcessDefinitionKey();
+    defaultUserClient
+        .newDeployResourceCommand()
+        .addProcessModel(
+            Bpmn.createExecutableProcess(PROCESS_ID)
+                .startEvent()
+                .userTask(USER_TASK_ID)
+                .zeebeUserTask()
+                .endEvent()
+                .done(),
+            "process.xml")
+        .send()
+        .join();
   }
 
   @Test
-  void shouldBeAuthorizedToMigrateProcessInstanceWithDefaultUser() {
+  void shouldBeAuthorizedToUpdateUserTaskWithDefaultUser() {
     // given
-    final var processInstanceKey =
-        defaultUserClient
-            .newCreateInstanceCommand()
-            .bpmnProcessId(PROCESS_ID)
-            .latestVersion()
-            .send()
-            .join()
-            .getProcessInstanceKey();
+    final var processInstanceKey = createProcessInstance();
+    final var userTaskKey = getUserTaskKey(processInstanceKey);
 
-    // when migrate to a non-existing process as authorization checks should fail first
-    // then
+    // when then
     final var response =
-        defaultUserClient
-            .newMigrateProcessInstanceCommand(processInstanceKey)
-            .migrationPlan(targetProcDefKey)
-            .addMappingInstruction(SOURCE_TASK, TARGET_TASK)
-            .send()
-            .join();
+        defaultUserClient.newUserTaskUpdateCommand(userTaskKey).priority(100).send().join();
 
     // The Rest API returns a null future for an empty response
     // We can verify for null, as if we'd be unauthenticated we'd get an exception
@@ -123,16 +96,10 @@ public class ProcessInstanceMigrationMigrateAuthorizationIT {
   }
 
   @Test
-  void shouldBeAuthorizedToMigrateProcessInstanceWithUser() {
+  void shouldBeAuthorizedToUpdateUserTaskWithUser() {
     // given
-    final var processInstanceKey =
-        defaultUserClient
-            .newCreateInstanceCommand()
-            .bpmnProcessId(PROCESS_ID)
-            .latestVersion()
-            .send()
-            .join()
-            .getProcessInstanceKey();
+    final var processInstanceKey = createProcessInstance();
+    final var userTaskKey = getUserTaskKey(processInstanceKey);
     final var username = UUID.randomUUID().toString();
     final var password = "password";
     authUtil.createUserWithPermissions(
@@ -140,19 +107,12 @@ public class ProcessInstanceMigrationMigrateAuthorizationIT {
         password,
         new Permissions(
             ResourceTypeEnum.PROCESS_DEFINITION,
-            PermissionTypeEnum.UPDATE_PROCESS_INSTANCE,
+            PermissionTypeEnum.UPDATE_USER_TASK,
             List.of(PROCESS_ID)));
 
     try (final var client = authUtil.createClient(username, password)) {
-      // when migrate to a non-existing process as authorization checks should fail first
-      // then
-      final var response =
-          client
-              .newMigrateProcessInstanceCommand(processInstanceKey)
-              .migrationPlan(targetProcDefKey)
-              .addMappingInstruction(SOURCE_TASK, TARGET_TASK)
-              .send()
-              .join();
+      // when then
+      final var response = client.newUserTaskUpdateCommand(userTaskKey).priority(100).send().join();
 
       // The Rest API returns a null future for an empty response
       // We can verify for null, as if we'd be unauthenticated we'd get an exception
@@ -161,29 +121,18 @@ public class ProcessInstanceMigrationMigrateAuthorizationIT {
   }
 
   @Test
-  void shouldBeUnauthorizedToMigrateProcessInstanceIfNoPermissions() {
+  void shouldBeUnauthorizedToUpdateUserTaskIfNoPermissions() {
     // given
-    final var processInstanceKey =
-        defaultUserClient
-            .newCreateInstanceCommand()
-            .bpmnProcessId(PROCESS_ID)
-            .latestVersion()
-            .send()
-            .join()
-            .getProcessInstanceKey();
+    final var processInstanceKey = createProcessInstance();
+    final var userTaskKey = getUserTaskKey(processInstanceKey);
     final var username = UUID.randomUUID().toString();
     final var password = "password";
     authUtil.createUser(username, password);
 
+    // when
     try (final var client = authUtil.createClient(username, password)) {
-      // when migrate to a non-existing process as authorization checks should fail first
-      // then
-      final var response =
-          client
-              .newMigrateProcessInstanceCommand(processInstanceKey)
-              .migrationPlan(targetProcDefKey)
-              .addMappingInstruction(SOURCE_TASK, TARGET_TASK)
-              .send();
+      // when we use the unauthorized client
+      final var response = client.newUserTaskUpdateCommand(userTaskKey).priority(100).send();
 
       // then
       assertThatThrownBy(response::join)
@@ -191,8 +140,28 @@ public class ProcessInstanceMigrationMigrateAuthorizationIT {
           .hasMessageContaining("title: FORBIDDEN")
           .hasMessageContaining("status: 403")
           .hasMessageContaining(
-              "Insufficient permissions to perform operation 'UPDATE_PROCESS_INSTANCE' on resource 'PROCESS_DEFINITION', required resource identifiers are one of '[*, %s]'",
+              "Insufficient permissions to perform operation 'UPDATE_USER_TASK' on resource 'PROCESS_DEFINITION', required resource identifiers are one of '[*, %s]'",
               PROCESS_ID);
     }
+  }
+
+  private long createProcessInstance() {
+    return defaultUserClient
+        .newCreateInstanceCommand()
+        .bpmnProcessId(PROCESS_ID)
+        .latestVersion()
+        .send()
+        .join()
+        .getProcessInstanceKey();
+  }
+
+  private static long getUserTaskKey(final long processInstanceKey) {
+    return RecordingExporter.userTaskRecords(UserTaskIntent.CREATED)
+        .withProcessInstanceKey(processInstanceKey)
+        .withElementId(USER_TASK_ID)
+        .limit(1)
+        .findFirst()
+        .orElseThrow()
+        .getKey();
   }
 }
