+++ zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/authorization/permissions/VariableDocumentUpdateAuthorizationTest.java
@@ -19,15 +19,14 @@ import io.camunda.client.protocol.rest.ResourceTypeEnum;
 import io.camunda.zeebe.it.util.AuthorizationsUtil;
 import io.camunda.zeebe.it.util.AuthorizationsUtil.Permissions;
 import io.camunda.zeebe.model.bpmn.Bpmn;
-import io.camunda.zeebe.protocol.record.intent.UserTaskIntent;
 import io.camunda.zeebe.qa.util.cluster.TestStandaloneBroker;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration;
 import io.camunda.zeebe.qa.util.junit.ZeebeIntegration.TestZeebe;
 import io.camunda.zeebe.test.util.junit.AutoCloseResources;
 import io.camunda.zeebe.test.util.junit.AutoCloseResources.AutoCloseResource;
-import io.camunda.zeebe.test.util.record.RecordingExporter;
 import io.camunda.zeebe.test.util.testcontainers.TestSearchContainers;
 import java.util.List;
+import java.util.Map;
 import java.util.UUID;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -38,9 +37,7 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 @AutoCloseResources
 @Testcontainers
 @ZeebeIntegration
-public class UserTaskAssignAuthorizationIT {
-  public static final String USER_TASK_ID = "userTask";
-
+public class VariableDocumentUpdateAuthorizationTest {
   @Container
   private static final ElasticsearchContainer CONTAINER =
       TestSearchContainers.createDefeaultElasticsearchContainer();
@@ -71,8 +68,7 @@ public class UserTaskAssignAuthorizationIT {
         .addProcessModel(
             Bpmn.createExecutableProcess(PROCESS_ID)
                 .startEvent()
-                .userTask(USER_TASK_ID)
-                .zeebeUserTask()
+                .userTask("task")
                 .endEvent()
                 .done(),
             "process.xml")
@@ -81,17 +77,15 @@ public class UserTaskAssignAuthorizationIT {
   }
 
   @Test
-  void shouldBeAuthorizedToAssignUserTaskWithDefaultUser() {
+  void shouldBeAuthorizedToUpdateVariablesWithDefaultUser() {
     // given
     final var processInstanceKey = createProcessInstance();
-    final var userTaskKey = getUserTaskKey(processInstanceKey);
 
     // when then
     final var response =
         defaultUserClient
-            .newUserTaskAssignCommand(userTaskKey)
-            .assignee("assignee")
-            .allowOverride(true)
+            .newSetVariablesCommand(processInstanceKey)
+            .variables(Map.of("foo", "bar"))
             .send()
             .join();
 
@@ -101,10 +95,9 @@ public class UserTaskAssignAuthorizationIT {
   }
 
   @Test
-  void shouldBeAuthorizedToAssignUserTaskWithUser() {
+  void shouldBeAuthorizedToUpdateVariablesWithUser() {
     // given
     final var processInstanceKey = createProcessInstance();
-    final var userTaskKey = getUserTaskKey(processInstanceKey);
     final var username = UUID.randomUUID().toString();
     final var password = "password";
     authUtil.createUserWithPermissions(
@@ -112,16 +105,15 @@ public class UserTaskAssignAuthorizationIT {
         password,
         new Permissions(
             ResourceTypeEnum.PROCESS_DEFINITION,
-            PermissionTypeEnum.UPDATE_USER_TASK,
+            PermissionTypeEnum.UPDATE_PROCESS_INSTANCE,
             List.of(PROCESS_ID)));
 
     try (final var client = authUtil.createClient(username, password)) {
       // when then
       final var response =
           client
-              .newUserTaskAssignCommand(userTaskKey)
-              .assignee("assignee")
-              .allowOverride(true)
+              .newSetVariablesCommand(processInstanceKey)
+              .variables(Map.of("foo", "bar"))
               .send()
               .join();
 
@@ -132,22 +124,18 @@ public class UserTaskAssignAuthorizationIT {
   }
 
   @Test
-  void shouldBeUnauthorizedToAssignUserTaskIfNoPermissions() {
+  void shouldBeUnauthorizedToUpdateVariablesIfNoPermissions() {
     // given
     final var processInstanceKey = createProcessInstance();
-    final var userTaskKey = getUserTaskKey(processInstanceKey);
     final var username = UUID.randomUUID().toString();
     final var password = "password";
     authUtil.createUser(username, password);
 
+    // when
     try (final var client = authUtil.createClient(username, password)) {
-      // when we use the unauthorized client
+      // when
       final var response =
-          client
-              .newUserTaskAssignCommand(userTaskKey)
-              .assignee("assignee")
-              .allowOverride(true)
-              .send();
+          client.newSetVariablesCommand(processInstanceKey).variables(Map.of("foo", "bar")).send();
 
       // then
       assertThatThrownBy(response::join)
@@ -155,7 +143,7 @@ public class UserTaskAssignAuthorizationIT {
           .hasMessageContaining("title: FORBIDDEN")
           .hasMessageContaining("status: 403")
           .hasMessageContaining(
-              "Insufficient permissions to perform operation 'UPDATE_USER_TASK' on resource 'PROCESS_DEFINITION', required resource identifiers are one of '[*, %s]'",
+              "Insufficient permissions to perform operation 'UPDATE_PROCESS_INSTANCE' on resource 'PROCESS_DEFINITION', required resource identifiers are one of '[*, %s]'",
               PROCESS_ID);
     }
   }
@@ -168,15 +156,5 @@ public class UserTaskAssignAuthorizationIT {
         .send()
         .join()
         .getProcessInstanceKey();
-  }
-
-  private static long getUserTaskKey(final long processInstanceKey) {
-    return RecordingExporter.userTaskRecords(UserTaskIntent.CREATED)
-        .withProcessInstanceKey(processInstanceKey)
-        .withElementId(USER_TASK_ID)
-        .limit(1)
-        .findFirst()
-        .orElseThrow()
-        .getKey();
   }
 }
